common: Sync extract-utils with template

+++ system/extract-files.sh
@@ -1,9 +1,8 @@
 #!/bin/bash
 #
-# Copyright (C) 2016 The CyanogenMod Project
-# Copyright (C) 2017-2020 The LineageOS Project
-# Copyright (C) 2020 Paranoid Android
-#
+# SPDX-FileCopyrightText: 2016 The CyanogenMod Project
+# SPDX-FileCopyrightText: 2017-2024 The LineageOS Project
+# SPDX-FileCopyrightText: 2020 Paranoid Android
 # SPDX-License-Identifier: Apache-2.0
 #
 
@@ -14,6 +13,10 @@ MY_DIR="${BASH_SOURCE%/*}"
 if [[ ! -d "${MY_DIR}" ]]; then MY_DIR="${PWD}"; fi
 
 ANDROID_ROOT="${MY_DIR}/../../../../"
+
+# If XML files don't have comments before the XML header, use this flag
+# Can still be used with broken XML files by using blob_fixup
+export TARGET_DISABLE_XML_FIXING=true
 
 HELPER="${ANDROID_ROOT}/tools/extract-utils/extract_utils.sh"
 if [ ! -f "${HELPER}" ]; then
@@ -54,19 +57,31 @@ fi
 function blob_fixup() {
     case "${1}" in
         product/etc/permissions/vendor.qti.hardware.data.connection-V1.0-java.xml | product/etc/permissions/vendor.qti.hardware.data.connection-V1.1-java.xml | product/etc/permissions/vendor.qti.hardware.data.connectionaidl-V1-java.xml)
+            [ "$2" = "" ] && return 0
             sed -i 's/xml version="2.0"/xml version="1.0"/g' "${2}"
             ;;
 
         system/framework/QXPerformance.jar)
+            [ "$2" = "" ] && return 0
             mv "${2}" "${2}.tmp"
             zipalign -p -f 4 "${2}.tmp" "${2}"
             rm "${2}.tmp"
             ;;
 
         system_ext/lib/libwfdservice.so | system_ext/lib64/libwfdservice.so)
+            [ "$2" = "" ] && return 0
             "${PATCHELF}" --replace-needed "android.media.audio.common.types-V2-cpp.so" "android.media.audio.common.types-V3-cpp.so" "${2}"
             ;;
+        *)
+            return 1
+            ;;
     esac
+
+    return 0
+}
+
+function blob_fixup_dry() {
+    blob_fixup "$1" ""
 }
 
 # Initialize the helper
